Add updateBounds and updateCosts to ObstacleLayer
Use lambda functions for Odometry subscription
Add `touch`, `worldToMap` and other utility functions to Layer base class as inline functions
Fix parameter passing to overloaded `std::bind` instead of using `boost::bind`
`updateCosts` and `updateBounds` use double for ranges instead of uint
Move MapCell array to Layer basecall

+++ costmap/include/costmap/costmap.h
@@ -28,7 +28,7 @@ class Costmap : public rclcpp::Node {
   void loadPlugin(std::shared_ptr<Layer> plugin);
 
   std::string global_frame_, base_frame_;
-  unsigned int minx_, miny_, maxx_, maxy_;
+  double minx_, miny_, maxx_, maxy_;
   int origin_x_, origin_y_;
 
   geometry_msgs::msg::Pose map_origin_;
